Update typeof
Using typeof is a great way to understand what type the value. EX: Lets say you have a stored variable and are unsure of the variables type.

Now as practice we can define seven new variables and set all seven variables to one variable to get the total length of this new variable.

We are also introduced to concatenation. Concatenation in JavaScript is the process of combining two or more strings together to form a single string. It is denoted by the '+' operator. For example, the following code concatenates the strings "Hello" and "World" together to form the string "HelloWorld":